// src/queue/generated/handlers/IServiceHandler.ts
/*
 * Copyright (c) Microsoft Corporation. All rights reserved.
 * Licensed under the MIT License. See License.txt in the project root for
 * license information.
 *
 * Code generated by Microsoft (R) AutoRest Code Generator.
 * Changes may cause incorrect behavior and will be lost if the code is
 * regenerated.
 */
// tslint:disable:max-line-length

import * as Models from "../artifacts/models";
import Context from "../Context";

export default interface IServiceHandler {
  setProperties(storageServiceProperties: Models.StorageServiceProperties, options: Models.ServiceSetPropertiesOptionalParams, context: Context): Promise<Models.ServiceSetPropertiesResponse>;
  getProperties(options: Models.ServiceGetPropertiesOptionalParams, context: Context): Promise<Models.ServiceGetPropertiesResponse>;
  getStatistics(options: Models.ServiceGetStatisticsOptionalParams, context: Context): Promise<Models.ServiceGetStatisticsResponse>;
  listQueuesSegment(options: Models.ServiceListQueuesSegmentOptionalParams, context: Context): Promise<Models.ServiceListQueuesSegmentResponse>;
}